initial refactors -> moved chat to AskQuestionForm
